[ADD]: 분반 진행률 collectionView extension 파일 분리
